I still can't get a hold of a sane way to arrange go code.
I want to structure a repo in such a way that git submodules can work, and both the primary repo and other projects that link it as a submodule have sane package paths (there's absolutely no sane reason why they should be different paths, period).  And I also want `go get` to work (though I myself will absolutely never use it, because the core concept of "pull whatever master is at this moment from somewhere on the internet and then run it" is unacceptable as an axiom).

I don't know why having my cake and eating it too is so difficult, but this directory layout, then symlinks in an subproject that includes it, is the only way I can get git submodules to work and have the same package path in this repo and the one that links this.

And then that breaks `go get` when used in a project that links to this one, because that project has to have a src/ dir, and this one already has that too, so now your package import names would look like crap again.  You can heap on more symlinks (i.e., put this as a submodule / go-get artifact at "src/polydawn.net/gosh" because that's what the project is called and the src dir is implicit if you make GOPATH inside your project, then symlink "src/polydawn.net/gosh/psh/"->"src/polydawn.net/gosh/src/polydawn.net/gosh/psh/")... but can you honestly tell me those make any sense to you from an objective standpoint, if you hadn't contrived this as a workaround to bizzare behavior restrictions by an overzealous toolchain?