Fix limitation that GetExitCode blocks even from inside exit listeners.
I think as a general pattern, it should be reasonable that all property and state retrieval methods on a command will return if that command is finished or panicked, without waiting.  Wait() waits, and it's obvious you shouldn't do that from inside an exit listener, and I still firmly want that to return only after exit listeners are done.  Retrieval methods that take timeouts will also still wait.

I wish I knew what the golang equivalant to java's volatile keyword was.  The golang race detector hasn't gone off on me yet with the change, but I feel like it's because I'm not pushing it hard enough.