ReaderFromInterface and WriterFromInterface now return ReadCloser and WriteCloser if possible.  Those methods are now documented with what types can be refined.
WriterFromInterface can produce a WriteCloser from both chan and chan<- types, which is entirely reasonable.  ReaderFromInterface returns ReadClosers for chan types, but Reader for <-chan types, which is entirely disgusting.  I have no idea why golang decided that the read size of a channel shouldn't support a close operation.  The fact that a ReadCloser interface exist seems like a clear enough argument to me that close(<-chan) ought to exist as well, if for no other reason than parity.